% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{\%>\%}
\alias{\%>\%}
\alias{chain_query}
\title{Pipe operator for chaining together operations.}
\description{
Imported from magrittr, use this to chain together
operations in a natural way.
}
\examples{
\dontrun{
# instead of
a(b(c("hello")), "bob")

# we can write:

c("hello") \\\%>\\\% b() \\\%>\\\% a("bob")

# this also allows for currying common arguments:

my_query <-
   fda_query("/drug/event.json") \\\%>\\\%
   fda_api_key("ABC")
}
}

