% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_api_key}
\alias{fda_api_key}
\title{Attach an API key to the query.}
\usage{
fda_api_key(q, key)
}
\value{
fda_query
}
\description{
All requests made to the API will have this API key
specified.
}

