% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_count}
\alias{fda_count}
\title{Count results from a given field.}
\usage{
fda_count(q, field)
}
\value{
fda_query
}
\description{
Returns a new query which will count the given field when
handed to \code{fda_exec}
}

