% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_debug}
\alias{fda_debug}
\title{Turn off/on API debugging.}
\usage{
fda_debug(on)
}
\description{
When set to TRUE, this will print additional debugging information from
the API, such as URLs being fetched.
}

