% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_exec}
\alias{fda_exec}
\title{Execute a query.}
\usage{
fda_exec(q)
}
\value{
data.frame containing API results
}
\description{
This actually sends a request to the OpenFDA API servers
which incorporates any filters or keys specified in the query
and returns the result as a data frame.
}

