% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_fetch}
\alias{fda_fetch}
\title{Fetch the given URL as JSON.}
\usage{
fda_fetch(url, catch_errors = TRUE)
}
\value{
data.frame
}
\description{
This uses jsonlite to fetch the URL.  The result is coerced into
a data frame.
}

