% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_filter}
\alias{fda_filter}
\title{Apply a filter to a query}
\usage{
fda_filter(q, name, value)
}
\arguments{
\item{name}{The field to filter on (e.g. "patient.patientonsetage")}

\item{value}{A value or range to filter on (e.g. "[1+TO+40]")}
}
\value{
fda_query
}
\description{
Apply a filter to a query
}

