% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_limit}
\alias{fda_limit}
\title{Set the number of results desired by this query.}
\usage{
fda_limit(q, limit)
}
\value{
fda_query
}
\description{
The default number of results returned by the API
is 1 (for search requests) and 100 (for count requests).
}

