% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_plot}
\alias{fda_plot}
\title{Plot a count query.}
\usage{
fda_plot(df)
}
\value{
ggplot2
}
\description{
This is just a convenience function which takes the
data frame output from fda_exec and gives it to qplot.
}

