% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_query}
\alias{fda_query}
\title{Create a new query.}
\usage{
fda_query(base)
}
\arguments{
\item{base}{The API endpoint to query: \itemize{
  \item /drug/event.json
  \item /drug/enforcement.json
  \item /device/event.json
  \item /food/event.json
 }}
}
\value{
fda_query
}
\description{
Queries can be extended with filters (using fda_filter).
They should be terminated using either \code{search} or
\code{count(field)}.
}

