% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_search}
\alias{fda_search}
\title{Fetch search results.}
\usage{
fda_search(q, field = FALSE)
}
\arguments{
\item{field}{The field to extract.  This should be a
dot (.) delimited path to follow to extract a field:
e.g. 'a.b.c.d', would perform the equivalent of
running a[[b]][[c]][[d]]}
}
\value{
fda_query
}
\description{
When combined with \code{fda_exec}, this query will
return a data frame consisting of the matching records.
}

