% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_skip}
\alias{fda_skip}
\title{Set the number of records to skip that match the search parameter.}
\usage{
fda_skip(q, skip)
}
\value{
fda_skip
}
\description{
Skip this number of records that match the search parameter, then return the matching records that follow.
Use in combination with limit to paginate results.
}
