% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fda_url}
\alias{fda_url}
\title{Return the URL that will be fetched for the current query.}
\usage{
fda_url(q)
}
\value{
character The url that will be fetched
}
\description{
Return the URL that will be fetched for the current query.
}

