% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\name{fetch_url}
\alias{fetch_url}
\title{Fetch the given URL as JSON.}
\usage{
fetch_url(url)
}
\value{
data.frame
}
\description{
This uses jsonlite to fetch the URL.  The result is coerced into
a data frame.
}

