% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFDA-package.R
\docType{package}
\name{openFDA-package}
\alias{openFDA-package}
\title{openFDA: 'openFDA' API}
\description{
The 'openFDA' API facilitates access to U.S. Food and Drug Administration (FDA) data on drugs, devices, foodstuffs, tobacco, and more. This package makes the API easily accessible with 'httr2', returning objects which the user can convert to JSON data and parse. The API is described in: Kass-Hout TA, Xu Z, Mohebbi M et al. (2016) \doi{10.1093/jamia/ocv153}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/simpar1471/openFDA}
  \item \url{https://simpar1471.github.io/openFDA/}
  \item Report bugs at \url{https://github.com/simpar1471/openFDA/issues}
}

}
\author{
\strong{Maintainer}: Simon Parker \email{simon.parker.24@ucl.ac.uk} (\href{https://orcid.org/0009-0003-8214-4496}{ORCID}) [copyright holder]

}
\keyword{internal}
