% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/openfda.R
\docType{package}
\name{openfda}
\alias{openfda}
\alias{openfda-package}
\title{openfda: A package for interfacing to the OpenFDA API}
\description{
This package provides a simple wrapper around the OpenFDA API.
}
\details{
It uses the \code{magrittr} piping interface to simplify
building complex queries.
}
\examples{
# Queries generally have the following format
fda_query(endpoint) \%>\%
   fda_filter(field, value) \%>\%
   fda_count(field) OR fda_search() \%>\%
   fda_exec()
}

