% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{openFDA_options}
\alias{openFDA_options}
\alias{current_openFDA_options}
\title{Get and set openFDA package-level options}
\usage{
openFDA_options(
  paging = NULL,
  paging_verbosity = NULL,
  handle_http_errors = NULL
)

current_openFDA_options()
}
\arguments{
\item{paging}{A single string used to set the \code{openFDA.paging} option. The
default, \code{NULL}, indicates that you don't want to change its value. The
\code{openFDA.paging} option controls how \code{openFDA()} behaves when there are
enough results that \href{https://open.fda.gov/apis/paging/}{paging is required}, and has the default value
\code{"ask"}. New values must be one of:
\itemize{
\item \code{"ask"} - \code{openFDA()} will warn you that pagination is required and ask
if you want this to be done. Depending on user input, either a single
\href{https://httr2.r-lib.org/}{httr2} response object or a list of
\href{https://httr2.r-lib.org/}{httr2} response objects will be returned. This
will throw an error if your R session is not in interactive mode. When s
etting this option, the value must be one of:
\item \code{"always"} - \code{openFDA()} will always perform pagination. A list of
\href{https://httr2.r-lib.org/}{httr2} responses will be returned when
pagination occurs.
\item \code{"never"} - \code{openFDA()} will never perform pagination. Only the first
\href{https://httr2.r-lib.org/}{httr2} response will be returned.
}}

\item{paging_verbosity}{A single string used to set the
\code{openFDA.paging_verbosity} option. The default, \code{NULL}, indicates that you
don't want to change its value. The \code{openFDA.paging_verbosity} option
controls whether \code{openFDA()} prints messages when paging is required, and
has the default value \code{"warn"}. New values must be one of:
\itemize{
\item \code{"verbose"} - If paging can be performed, print a message to the console,
stating how many requests are required with a minimum estimate for the
time this will take.
\item \code{"quiet"} - Print no messages to the console regarding paging.
}}

\item{handle_http_errors}{A single string used to set the
\code{openFDA.handle_http_errors} option. The default, \code{NULL}, indicates that
you don't want to change its value. The \code{openFDA.handle_http_errors} option
controls how \code{openFDA()} reacts when responses from the openFDA API have
HTTP codes other than \verb{200 OK}, and has the default value \code{"warn"}. New
values must be one of:
\itemize{
\item \code{"warn"}: If the returned HTTP code is not 200, issue a warning to the
console and return the underlying \href{https://httr2.r-lib.org/}{httr2}
response object.
\item \code{"error"}: If the returned HTTP code is not 200, throw an (informative)
error.
\item \code{"silent"}: If the returned HTTP code is not 200, return the underlying
\href{https://httr2.r-lib.org/}{httr2} response object without printing a
warning.
}}
}
\value{
For \code{openFDA_options()} returns a named character vector with
current option values. This will include the updated option values, if any
have been set.

For \code{current_openFDA_options()}, returns a named character
vector with the current values for each openFDA package-level option.
}
\description{
User-configurable options used by openFDA, which provide a
mechanism for setting default behaviours.
}
\examples{
# Print current options by leaving all input args as `NULL`:
openFDA_options()

# Or specify some options to set these:
openFDA_options(paging = "never", handle_http_errors = "silent")

# Make sure to give valid option values
try(openFDA_options(paging = "bad-value"))

# If you set an option using `options()` and it is not valid, you will
# invoke errors
try(
  rlang::with_options(openFDA.paging = "no", current_openFDA_options())
)
}
