% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFDA-package.R
\name{openFDA_options}
\alias{openFDA_options}
\title{Options consulted by openFDA}
\description{
User-configurable options used by openFDA, which provide a mechanism
for setting default behaviour.

If the default doesn't suit you, change the option using or specify the value
of \code{paging} in \code{openFDA()}.
}
\section{Options for the openFDA package}{

\itemize{
\item \code{openFDA.paging}: customize the default behaviour of \code{openFDA()} when there
are enough results that \href{https://open.fda.gov/apis/paging/}{paging is required}. The new value of this option
should be one of \code{c("ask", "yes", "no", "yes-quiet", "no-quiet")}.
}
}

\examples{
# Access with base R...
options("openFDA.paging")

# ... or use the rlang package
rlang::peek_options("openFDA.paging")

# Set using base R...
options(openFDA.paging = "yes")

# ... or use the rlang package
rlang::push_options(openFDA.paging = "no")

# (Just to check it's changed)
rlang::peek_options("openFDA.paging")
}
