% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openFDA-package.R
\name{openFDA_options}
\alias{openFDA_options}
\title{Options consulted by openFDA}
\value{
No return value - they're package-level options, which can be used to
change the behaviour of \code{openFDA()}.
}
\description{
User-configurable options used by openFDA, which provide a mechanism
for setting default behaviour.
}
\section{Options for the openFDA package}{

\itemize{
\item \code{openFDA.paging}: customise the default behaviour of \code{openFDA()} when there
are enough results that \href{https://open.fda.gov/apis/paging/}{paging is required}. The default is \code{"ask"}. When
setting this option, the value must be one of:
\itemize{
\item \code{"ask"} - \code{openFDA()} will warn you that pagination is required and ask
if you want this to be done. Depending on user input, either a single
\href{https://httr2.r-lib.org/}{httr2} response object or a list of
\href{https://httr2.r-lib.org/}{httr2} response objects will be returned. This
will throw an error if your R session is not in interactive mode. When s
etting this option, the value must be one of:
\item \code{"always"} - \code{openFDA()} will always perform pagination. A list of
\href{https://httr2.r-lib.org/}{httr2} responses will be returned when
pagination occurs.
\item \code{"never"} - \code{openFDA()} will never perform pagination. Only the first
\href{https://httr2.r-lib.org/}{httr2} response will be returned.
}
\item \code{openFDA.paging_verbosity}: customise openFDA's verbosity when paging. The
default is \code{"verbose"}. When setting this option, the value must be one of:
\itemize{
\item \code{"verbose"} - If paging can be performed, print a message to the console,
stating how many requests are required with a minimum estimate for the
time this will take.
\item \code{"quiet"} - Print no messages to the console regarding paging.
}
\item \code{openFDA.handle_http_errors}: customise whether \code{openFDA()} prints warnings
to the console when requests to the openFDA server were not \verb{200 OK}. The
default is \code{"warn"}. When setting this option, the value must be one of:
\itemize{
\item \code{"warn"}: If the returned HTTP code is not 200, issue a warning to the
console and return the underlying \href{https://httr2.r-lib.org/}{httr2}
response object.
\item \code{"error"}: If the returned HTTP code is not 200, throw an (informative)
error.
\item \code{"silent"}: If the returned HTTP code is not 200, return the underlying
\href{https://httr2.r-lib.org/}{httr2} response object without printing a
warning.
}
}
}

\examples{
# Access with base R...
options("openFDA.paging")

# ... or use the rlang package
rlang::peek_options("openFDA.paging")

# Set using base R...
options(openFDA.paging = "yes")

# ... or use the rlang package
rlang::push_options(openFDA.paging = "no")

# (Just to check it's changed)
rlang::peek_option("openFDA.paging")
}
